+++ .cargo/config.toml
@@ -4,3 +4,6 @@
 # we don't use `[build]` because of rust analyzer's build cache invalidation https://github.com/emilk/eframe_template/issues/93
 [target.wasm32-unknown-unknown]
 rustflags = ["--cfg=web_sys_unstable_apis"]
+
+[build]
+target = "x86_64-pc-windows-gnu"
