+++ Cargo.toml
@@ -12,7 +12,7 @@ eframe = { version = "0.23.0", default-features = false, features = [
     #"accesskit",     # Make egui comptaible with screen readers. NOTE: adds a lot of dependencies.
     "default_fonts", # Embed the default egui fonts.
     "glow",          # Use the glow rendering backend. Alternative: "wgpu".
-    "persistence",   # Enable restoring app state when restarting the app.
+    # "persistence",   # Enable restoring app state when restarting the app.
 ] }
 log = "0.4"
