+++ win.memo
@@ -1,0 +1,6 @@
+# memo on account of working under wsl2
+#debug
+#cross build --target x86_64-pc-windows-gnu
+
+#release
+cross build --release --target x86_64-pc-windows-gnu
